[FEAT] #45 - 지갑 입금 기능 추가

+++ src/main/java/com/thock/back/api/boundedContext/payment/app/PaymentSupport.java
@@ -27,7 +27,7 @@ public class PaymentSupport {
     }
 
     @Transactional(readOnly = true)
-    public Optional<Wallet> findWalletByHolderId(int holderId) {
+    public Optional<Wallet> findWalletByHolderId(Long holderId) {
         return walletRepository.findByHolderId(holderId);
     }
